integration of error and success component & removing info
