--- src/app/modules/hint/components/hint.component.html
@@ -1,17 +1,0 @@
-<div *ngIf="visible">
-  <ng-container *ngIf="errors$">
-    <div *ngFor="let error of errors$ | async" class="hint hint--error">
-      <p>{{ error?.message }}</p>
-    </div>
-  </ng-container>
-  <ng-container *ngIf="piecesOfInformation$">
-  <div *ngFor="let info of piecesOfInformation$ | async" class="hint hint--info">
-    <p>{{ info }}</p>
-  </div>
-</ng-container>
-  <ng-container *ngIf="successes$">
-  <div *ngFor="let success of successes$ | async" class="hint hint--success">
-    <p>{{ success.success }}</p>
-  </div>
-</ng-container>
-</div>
