+++ src/app/modules/hint/components/hint/hint.component.scss
@@ -1,15 +1,18 @@
 @import '~src/styles/variables';
-
-.hint {
+.hint-container {
   position: absolute;
   bottom:5em;
   right: 3em;
+}
+
+.hint {
   background-color: white;
   width: 15vw;
   border: 3px solid;
   border-radius: $border-radius;
   box-shadow: $box-shadow;
   padding: 1em;
+  margin: .5em;
 
   &--success {
     @include setColor($color-successful);
