+++ src/app/modules/hint/components/hint/hint.component.spec.ts
@@ -1,6 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
 import { HintComponent } from './hint.component';
+
 
 describe('HintComponent', () => {
   let component: HintComponent;
